requirements.txt
Flask==1.1.2
Jinja2==2.11.2
MarkupSafe==1.1.1
Werkzeug==1.0.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
idna==2.10
itsdangerous==1.1.0
joblib==0.16.0
numpy==1.19.2
opencv-python-headless==4.5.1.48
opencv-contrib-python-headless==4.5.1.48
tensorflow-cpu==2.5.0
pip==20.2.3
requests==2.24.0
scipy==1.5.2
setuptools==41.2.0
threadpoolctl==2.1.0
urllib3==1.25.10
wheel==0.35
gunicorn==20.0.4